update module dependencies
