+++ apps/server/src/modules/board/board.module.ts
@@ -7,6 +7,7 @@ import { ToolConfigModule } from '@modules/tool/tool-config.module';
 import { UserModule } from '@modules/user/user.module';
 import { HttpModule } from '@nestjs/axios';
 import { Module } from '@nestjs/common';
+import { CqrsModule } from '@nestjs/cqrs';
 import { ContentElementFactory } from '@shared/domain/domainobject';
 import { CourseRepo } from '@shared/repo';
 import { LoggerModule } from '@src/core/logger';
@@ -22,6 +23,7 @@ import {
 	MediaElementService,
 	MediaLineService,
 	SubmissionItemService,
+	UserDeletedEventHandlerService,
 } from './service';
 import { BoardDoCopyService, SchoolSpecificFileCopyServiceFactory } from './service/board-do-copy-service';
 import { ColumnBoardCopyService } from './service/column-board-copy.service';
@@ -37,6 +39,7 @@ import { ColumnBoardCopyService } from './service/column-board-copy.service';
 		HttpModule,
 		ToolConfigModule,
 		TldrawClientModule,
+		CqrsModule,
 	],
 	providers: [
 		BoardDoAuthorizableService,
@@ -58,6 +61,7 @@ import { ColumnBoardCopyService } from './service/column-board-copy.service';
 		MediaBoardService,
 		MediaLineService,
 		MediaElementService,
+		UserDeletedEventHandlerService,
 	],
 	exports: [
 		BoardDoAuthorizableService,
