+++ apps/server/src/modules/board/service/event/index.ts
@@ -1,0 +1,1 @@
+export { UserDeletedEventHandlerService } from './user-deleted-event-handler.service';
